Create labels file and relocate labels there
