+++ src/js/edu-benefits/1995/containers/ConfirmationPage.jsx
@@ -5,7 +5,7 @@ import Scroll from 'react-scroll';
 
 import { focusElement } from '../../../common/utils/helpers';
 
-import { benefitsLabels } from '../../utils/helpers';
+import { benefitsLabels } from '../../utils/labels';
 
 const scroller = Scroll.scroller;
 const scrollToTop = () => {
